🎨Increases color variety and fixes some aspects of debounce useage🎨
I originally got rid of the intermediate qrlink value that would then
have 'https://' appended to it when the update button was clicked. I
had it added before form submission, so it was still valid and would
save the correct value on the backend. However, I missed that the qrcode
that was getting generated on the frontend did not have 'https://' on
the front of it. Technically, this would still work, as browsers would
add that automatically. However, it would result in the QRCode being
shown during generation and the QRCode getting displayed in the list of
QRCodes afterwards being different (as one would have the https on the
front and one wouldn't). Therefore, I changed the form.url to be a computed
value that was either an empty string (if qrlink is an empty string) or
https:// + qrlink.value. If I always added the https on the front then the
form wouldn't hide properly when there was nothing in the input.