🧽Cleans up the debounce function and corresponding types🧽
Changed DebouncedFunction to actually be the function itself and
the cancel function, and then set DebouncedFunction as the return
type of debounce, set what was teardown to debouncedFunc.cancel,
and then just return debouncedFunc instead of an array of that and
the teardown. I also fixed up the usage of it on the DebouncedInput
component to reflect this change.